feat: Add environmental analysis and ML model status sensors, along with training and analysis services

+++ custom_components/area_occupancy/ml_models.py
@@ -252,7 +252,7 @@ class MLModelManager:
             if not X:
                 return None, None
 
-            return np.array(X), np.array(y)
+            return np.array(X, dtype=np.float64), np.array(y, dtype=np.int64)
 
         except Exception as err:
             _LOGGER.error("Failed to prepare training data: %s", err)
